refactor: remove duplicated code
