+++ sdk/src/l2/mod.rs
@@ -1,5 +1,1 @@
 pub mod constants;
-pub mod deposit;
-pub mod l1_to_l2_tx_data;
-pub mod merkle_tree;
-pub mod withdraw;
