+++ sdk/src/sdk.rs
@@ -1,16 +1,21 @@
-use crate::client::{EthClient, EthClientError, Overrides};
-use ethrex_common::types::GenericTransaction;
-use ethrex_common::{Address, H256, U256};
-use ethrex_l2_rpc::signer::{LocalSigner, Signer};
-use ethrex_rpc::types::receipt::RpcReceipt;
+use ethrex_common::{Address, U256, types::TxType};
+use ethrex_l2_rpc::{
+    clients::send_generic_transaction,
+    signer::{LocalSigner, Signer},
+};
+use ethrex_rpc::{
+    EthClient,
+    clients::{EthClientError, Overrides},
+    types::receipt::RpcReceipt,
+};
+use ethrex_sdk::{calldata::encode_calldata, get_address_from_secret_key};
+use keccak_hash::H256;
 use secp256k1::SecretKey;
 
-pub mod calldata;
 pub mod client;
 pub mod create;
 pub mod errors;
 pub mod keystore;
-pub mod privileged_transaction_data;
 pub mod sign;
 pub mod utils;
 
@@ -22,43 +27,6 @@ pub enum SdkError {
     FailedToParseAddressFromHex,
     #[error("Failed deserializing log: {0}")]
     FailedToDeserializeLog(String),
-}
-
-pub async fn transfer(
-    amount: U256,
-    from: Address,
-    to: Address,
-    private_key: &SecretKey,
-    client: &EthClient,
-) -> Result<H256, EthClientError> {
-    let gas_price = client
-        .get_gas_price_with_extra(20)
-        .await?
-        .try_into()
-        .map_err(|_| {
-            EthClientError::InternalError("Failed to convert gas_price to a u64".to_owned())
-        })?;
-
-    let mut tx = client
-        .build_eip1559_transaction(
-            to,
-            from,
-            Default::default(),
-            Overrides {
-                value: Some(amount),
-                max_fee_per_gas: Some(gas_price),
-                max_priority_fee_per_gas: Some(gas_price),
-                ..Default::default()
-            },
-        )
-        .await?;
-
-    let mut tx_generic: GenericTransaction = tx.clone().into();
-    tx_generic.from = from;
-    let gas_limit = client.estimate_gas(tx_generic).await?;
-    tx.gas_limit = gas_limit;
-    let signer = Signer::Local(LocalSigner::new(*private_key));
-    client.send_eip1559_transaction(&tx, &signer).await
 }
 
 pub async fn wait_for_transaction_receipt(
@@ -117,6 +85,44 @@ pub fn balance_in_eth(eth: bool, balance: U256) -> String {
     } else {
         format!("{balance}")
     }
+}
+
+pub async fn deposit_through_contract_call(
+    amount: U256,
+    to: Address,
+    depositor_private_key: &SecretKey,
+    bridge_address: Address,
+    eth_client: &EthClient,
+) -> Result<H256, EthClientError> {
+    let l1_from = get_address_from_secret_key(depositor_private_key)?;
+    let calldata = encode_calldata("deposit(address)", &[Value::Address(to)])?;
+    let gas_price = eth_client
+        .get_gas_price_with_extra(20)
+        .await?
+        .try_into()
+        .map_err(|_| {
+            EthClientError::InternalError("Failed to convert gas_price to a u64".to_owned())
+        })?;
+
+    let deposit_tx = eth_client
+        .build_generic_tx(
+            TxType::EIP1559,
+            bridge_address,
+            l1_from,
+            calldata.into(),
+            Overrides {
+                from: Some(l1_from),
+                value: Some(amount),
+                max_fee_per_gas: Some(gas_price),
+                max_priority_fee_per_gas: Some(gas_price),
+                ..Default::default()
+            },
+        )
+        .await?;
+
+    let signer = Signer::Local(LocalSigner::new(*depositor_private_key));
+
+    send_generic_transaction(eth_client, deposit_tx, &signer).await
 }
 
 #[test]
